text fix

+++ Assets/Scripts/UI/UpgradeMenu/Upgrade.cs
@@ -20,15 +20,15 @@ public class Upgrade
         switch (upgradeType)
         {
             case UpgradeType.NormalDamagePercent:
-                return "Increase Normal Attack Damage by " + upgradeValue * 100 + "%";
+                return "Increase Normal Attack Damage by " + (int)(upgradeValue * 100) + "%";
             case UpgradeType.PierceDamagePercent:
-                return "Increase Pierce Attack Damage by " + upgradeValue * 100 + "%";
+                return "Increase Pierce Attack Damage by " + (int)(upgradeValue * 100) + "%";
             case UpgradeType.MaxManaPercent:
-                return "Increase Max Mana by " + upgradeValue * 100 + "%";
+                return "Increase Max Mana by " + (int)(upgradeValue * 100) + "%";
             case UpgradeType.MaxHealthPercent:
-                return "Increase Max Health by " + upgradeValue * 100 + "%";
+                return "Increase Max Health by " + (int)(upgradeValue * 100) + "%";
             case UpgradeType.AbilityCooldownPercent:
-                return "Decrease Ability Cooldown by " + upgradeValue * 100 + "%";
+                return "Decrease Ability Cooldown by " + (int)(upgradeValue * 100) + "%";
             default:
                 return "Unknown";
         }
